cleanups and moving of code
